+++ packages/react/src/components/elements/button.tsx
@@ -1,8 +1,18 @@
 import { ButtonProps, Button as ShadButton } from "../ui/button";
 import { Loader2 } from "lucide-react";
 
+export interface ButtonStyles {
+  root: React.CSSProperties;
+}
+
+export interface ButtonClassNames {
+  root: string;
+}
+
 interface Props extends ButtonProps {
   isLoading?: boolean;
+  styles?: Partial<ButtonStyles>;
+  classNames?: Partial<ButtonClassNames>;
 }
 
 export const Button = ({
@@ -10,10 +20,18 @@ export const Button = ({
   children,
   isLoading,
   disabled,
+  classNames,
+  styles,
   ...props
 }: Props) => {
   return (
-    <ShadButton id={id} disabled={disabled || isLoading} {...props}>
+    <ShadButton
+      id={id}
+      className={classNames?.root}
+      disabled={disabled || isLoading}
+      style={styles?.root}
+      {...props}
+    >
       {isLoading && (
         <>
           <Loader2 className="mr-2 h-4 w-4 animate-spin" />