+++ packages/react/src/components/elements/numberInput/stepper.tsx
@@ -2,6 +2,18 @@ import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { MinusIcon, PlusIcon } from "@radix-ui/react-icons";
 
+export interface NumberInputStepperStyles {
+  root: React.CSSProperties;
+  stepButton: React.CSSProperties;
+  valueText: React.CSSProperties;
+}
+
+export interface NumberInputStepperClassNames {
+  root: string;
+  stepButton: string;
+  valueText: string;
+}
+
 export interface NumberInputStepperProps {
   onChange?: (value: number) => unknown;
   value?: number;
@@ -14,6 +26,8 @@ export interface NumberInputStepperProps {
   isRequired?: boolean;
   disabled?: boolean;
   readOnly?: boolean;
+  styles?: Partial<NumberInputStepperStyles>;
+  classNames?: Partial<NumberInputStepperClassNames>;
 }
 
 export function NumberInputStepper({
@@ -22,6 +36,8 @@ export function NumberInputStepper({
   step = { increment: 1, decrement: 1 },
   disabled,
   readOnly,
+  classNames,
+  styles,
 }: NumberInputStepperProps): React.JSX.Element {
   return (
     <div
@@ -29,11 +45,16 @@ export function NumberInputStepper({
         "h-10 flex items-center gap-4",
         Boolean(disabled || readOnly) && "opacity-50",
         disabled && "cursor-not-allowed",
+        classNames?.root,
       )}
+      style={styles?.root}
     >
       <Button
         variant="outline"
-        className="h-7 w-7 bg-transparent p-0 opacity-50 hover:opacity-100"
+        className={cn(
+          "h-7 w-7 bg-transparent p-0 opacity-50 hover:opacity-100",
+          classNames?.stepButton,
+        )}
         onClick={() =>
           onChange && onChange((value ?? 0) - (step?.decrement ?? 1))
         }
@@ -42,10 +63,18 @@ export function NumberInputStepper({
       >
         <MinusIcon className="h-4 w-4" />
       </Button>
-      <p className="text-sm">{value}</p>
+      <p
+        className={cn("text-sm", classNames?.valueText)}
+        style={styles?.valueText}
+      >
+        {value}
+      </p>
       <Button
         variant="outline"
-        className="h-7 w-7 bg-transparent p-0 opacity-50 hover:opacity-100"
+        className={cn(
+          "h-7 w-7 bg-transparent p-0 opacity-50 hover:opacity-100",
+          classNames?.stepButton,
+        )}
         onClick={() =>
           onChange && onChange((value ?? 0) + (step?.increment ?? 1))
         }