+++ packages/react/src/components/elements/threeColumn.tsx
@@ -1,13 +1,35 @@
 import { cn } from "@/lib/utils";
 import React from "react";
 
-interface Props extends React.HTMLAttributes<HTMLDivElement> {}
+export interface ThreeColumnStyles {
+  root: React.CSSProperties;
+}
 
-export const ThreeColumn = ({ children, ...props }: Props) => {
+export interface ThreeColumnClassNames {
+  root: string;
+}
+
+interface Props extends React.HTMLAttributes<HTMLDivElement> {
+  styles?: Partial<ThreeColumnStyles>;
+  classNames?: Partial<ThreeColumnClassNames>;
+}
+
+export const ThreeColumn = ({
+  children,
+  classNames,
+  styles,
+  className,
+  ...props
+}: Props) => {
   return (
     <div
+      className={cn(
+        "w-full grid grid-cols-3 gap-2",
+        className,
+        classNames?.root,
+      )}
+      style={styles?.root}
       {...props}
-      className={cn(props?.className, "w-full grid grid-cols-3 gap-2")}
     >
       {children}
     </div>