+++ packages/types/src/elements/repeaterGroup/index.ts
@@ -4,13 +4,22 @@ import {
   GenericLezzformElement,
   LezzformElementCategory,
 } from "../..";
+import { GeneralStyleMetadata } from "../../styles";
 
 export type RepeaterGroupType = GenericLezzformElement<
-  GenericFieldAttributes & RepeaterGroupAttributesType
+  GenericFieldAttributes & RepeaterGroupAttributesType,
+  RepeaterGroupStyleType
 > & {
   type: "RepeaterGroup";
   category: LezzformElementCategory.Repeater;
   rule: RepeaterGroupRuleType;
+};
+
+export type RepeaterGroupStyleType = GeneralStyleMetadata;
+
+export type RepeaterGroupStyleMetadataValuesType = {
+  deleteButton: object;
+  insertButton: object;
 };
 
 export type RepeaterGroupRuleType = ArrayRule;