+++ packages/types/src/elements/twoColumn/index.ts
@@ -1,6 +1,9 @@
 import { GenericLezzformElement, LezzformElementCategory } from "../..";
+import { GeneralStyleMetadata } from "../../styles";
 
-export type TwoColumnType = GenericLezzformElement & {
+export type TwoColumnType = GenericLezzformElement<{}, TwoColumnStyleType> & {
   type: "TwoColumn";
   category: LezzformElementCategory.Layout;
 };
+
+export type TwoColumnStyleType = GeneralStyleMetadata;